requirements.txt
python-docx>=0.8.10
openpyxl>=3.0.7
requests>=2.25.1
reportlab>=3.6.0
fpdf>=1.7.2
PyPDF2>=1.26.0
babel>=2.9.0
pillow>=8.0.0 